guess_game-main/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
opentelemetry-instrumentation-flask
opentelemetry-distro
opentelemetry-distro[otlp]
